refactor forge stuff
