+++ platforms/forge/src/main/java/com/dfsek/terra/forge/TerraForgePlugin.java
@@ -37,11 +37,11 @@ import com.dfsek.terra.config.lang.LangUtil;
 import com.dfsek.terra.config.lang.Language;
 import com.dfsek.terra.config.pack.ConfigPack;
 import com.dfsek.terra.config.templates.BiomeTemplate;
+import com.dfsek.terra.forge.generation.ForgeChunkGeneratorWrapper;
+import com.dfsek.terra.forge.generation.PopulatorFeature;
+import com.dfsek.terra.forge.generation.TerraLevelType;
 import com.dfsek.terra.forge.handle.ForgeItemHandle;
 import com.dfsek.terra.forge.handle.ForgeWorldHandle;
-import com.dfsek.terra.forge.world.features.PopulatorFeature;
-import com.dfsek.terra.forge.world.generator.ForgeChunkGeneratorWrapper;
-import com.dfsek.terra.forge.world.generator.config.TerraLevelType;
 import com.dfsek.terra.profiler.Profiler;
 import com.dfsek.terra.profiler.ProfilerImpl;
 import com.dfsek.terra.registry.exception.DuplicateEntryException;
