+++ platforms/forge/src/main/java/com/dfsek/terra/forge/block/ForgeBlock.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.forge.world.block;
+package com.dfsek.terra.forge.block;
 
 import com.dfsek.terra.api.math.vector.Location;
 import com.dfsek.terra.api.platform.block.Block;
@@ -7,7 +7,7 @@ import com.dfsek.terra.api.platform.block.BlockFace;
 import com.dfsek.terra.api.platform.block.BlockType;
 import com.dfsek.terra.api.platform.block.state.BlockState;
 import com.dfsek.terra.api.platform.world.World;
-import com.dfsek.terra.forge.world.ForgeAdapter;
+import com.dfsek.terra.forge.ForgeAdapter;
 import net.minecraft.block.FlowingFluidBlock;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IWorld;
