+++ platforms/forge/src/main/java/com/dfsek/terra/forge/block/data/ForgeDirectional.java
@@ -1,9 +1,9 @@
-package com.dfsek.terra.forge.world.block.data;
+package com.dfsek.terra.forge.block.data;
 
 import com.dfsek.terra.api.platform.block.BlockFace;
 import com.dfsek.terra.api.platform.block.data.Directional;
-import com.dfsek.terra.forge.world.ForgeAdapter;
-import com.dfsek.terra.forge.world.block.ForgeBlockData;
+import com.dfsek.terra.forge.ForgeAdapter;
+import com.dfsek.terra.forge.block.ForgeBlockData;
 import net.minecraft.block.BlockState;
 import net.minecraft.state.DirectionProperty;
