+++ platforms/forge/src/main/java/com/dfsek/terra/forge/block/data/ForgeOrientable.java
@@ -1,12 +1,11 @@
-package com.dfsek.terra.forge.world.block.data;
+package com.dfsek.terra.forge.block.data;
 
 import com.dfsek.terra.api.platform.block.Axis;
 import com.dfsek.terra.api.platform.block.data.Orientable;
-import com.dfsek.terra.forge.world.block.ForgeBlockData;
+import com.dfsek.terra.forge.block.ForgeBlockData;
 import net.minecraft.block.BlockState;
 import net.minecraft.state.EnumProperty;
 import net.minecraft.util.Direction;
-
 
 import java.util.Arrays;
 import java.util.Set;
