+++ platforms/forge/src/main/java/com/dfsek/terra/forge/block/data/ForgeSlab.java
@@ -1,4 +1,4 @@
-package com.dfsek.terra.forge.world.block.data;
+package com.dfsek.terra.forge.block.data;
 
 import com.dfsek.terra.api.platform.block.data.Slab;
 import net.minecraft.block.BlockState;
