+++ platforms/forge/src/main/java/com/dfsek/terra/forge/block/data/ForgeWaterlogged.java
@@ -1,7 +1,7 @@
-package com.dfsek.terra.forge.world.block.data;
+package com.dfsek.terra.forge.block.data;
 
 import com.dfsek.terra.api.platform.block.data.Waterlogged;
-import com.dfsek.terra.forge.world.block.ForgeBlockData;
+import com.dfsek.terra.forge.block.ForgeBlockData;
 import net.minecraft.block.BlockState;
 import net.minecraft.state.properties.BlockStateProperties;
