+++ platforms/forge/src/main/java/com/dfsek/terra/forge/generation/PopulatorFeature.java
@@ -1,8 +1,7 @@
-package com.dfsek.terra.forge.world.features;
+package com.dfsek.terra.forge.generation;
 
 import com.dfsek.terra.api.platform.world.Chunk;
 import com.dfsek.terra.api.platform.world.World;
-import com.dfsek.terra.forge.world.generator.ForgeChunkGeneratorWrapper;
 import com.mojang.serialization.Codec;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.ISeedReader;
