+++ platforms/forge/src/main/java/com/dfsek/terra/forge/handle/ForgeWorldHandle.java
@@ -2,8 +2,8 @@ package com.dfsek.terra.forge.handle;
 
 import com.dfsek.terra.api.platform.entity.EntityType;
 import com.dfsek.terra.api.platform.handle.WorldHandle;
-import com.dfsek.terra.forge.world.ForgeAdapter;
-import com.dfsek.terra.forge.world.block.ForgeBlockData;
+import com.dfsek.terra.forge.ForgeAdapter;
+import com.dfsek.terra.forge.block.ForgeBlockData;
 import com.mojang.brigadier.StringReader;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import net.minecraft.block.BlockState;
