+++ platforms/forge/src/main/java/com/dfsek/terra/forge/mixin/implementations/block/BlockEntityMixin.java
@@ -3,8 +3,8 @@ package com.dfsek.terra.forge.mixin.implementations.block;
 import com.dfsek.terra.api.platform.block.Block;
 import com.dfsek.terra.api.platform.block.BlockData;
 import com.dfsek.terra.api.platform.block.state.BlockState;
-import com.dfsek.terra.forge.world.ForgeAdapter;
-import com.dfsek.terra.forge.world.block.ForgeBlock;
+import com.dfsek.terra.forge.ForgeAdapter;
+import com.dfsek.terra.forge.block.ForgeBlock;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
