+++ platforms/forge/src/main/java/com/dfsek/terra/forge/mixin/implementations/chunk/ChunkRegionMixin.java
@@ -4,8 +4,8 @@ import com.dfsek.terra.api.platform.block.Block;
 import com.dfsek.terra.api.platform.block.BlockData;
 import com.dfsek.terra.api.platform.world.Chunk;
 import com.dfsek.terra.api.platform.world.World;
-import com.dfsek.terra.forge.world.block.ForgeBlock;
-import com.dfsek.terra.forge.world.block.ForgeBlockData;
+import com.dfsek.terra.forge.block.ForgeBlock;
+import com.dfsek.terra.forge.block.ForgeBlockData;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.gen.WorldGenRegion;
 import org.jetbrains.annotations.NotNull;
