+++ platforms/forge/src/main/java/com/dfsek/terra/forge/mixin/implementations/world/ServerWorldMixin.java
@@ -9,8 +9,8 @@ import com.dfsek.terra.api.platform.world.World;
 import com.dfsek.terra.api.platform.world.generator.ChunkGenerator;
 import com.dfsek.terra.api.platform.world.generator.GeneratorWrapper;
 import com.dfsek.terra.api.world.generation.TerraChunkGenerator;
-import com.dfsek.terra.forge.world.block.ForgeBlock;
-import com.dfsek.terra.forge.world.generator.ForgeChunkGeneratorWrapper;
+import com.dfsek.terra.forge.block.ForgeBlock;
+import com.dfsek.terra.forge.generation.ForgeChunkGeneratorWrapper;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IServerWorld;
 import net.minecraft.world.server.ServerWorld;
